refactor: extract merchant distribution card
