+++ frontend/src/insights/InsightsPage.tsx
@@ -45,6 +45,10 @@ import {
 } from './components/IncomeExpenseBreakdownCard'
 import { InsightsRangeSelector, type InsightsRangeSelectorOption } from './components/InsightsRangeSelector'
 import {
+  MerchantDistributionCard,
+  type MerchantMarketTile,
+} from './components/MerchantDistributionCard'
+import {
   IncomeExpenseSankeyCard,
   type IncomeExpenseFlowData,
   type IncomeExpenseFlowNode,
@@ -92,21 +96,6 @@ type InsightScaffoldData = {
 }
 
 type NetWorthGranularity = 'day' | 'week' | 'month'
-
-type MerchantMarketTile = MerchantBubble & {
-  x: number
-  y: number
-  width: number
-  height: number
-  changePct: number
-  changeAmount: number | null
-}
-
-type MerchantMarketHover = {
-  merchant: MerchantMarketTile
-  x: number
-  y: number
-}
 
 type PeriodBrief = {
   metrics: Array<{
@@ -863,22 +852,6 @@ function getMerchantChangeAmount(totalAmount: number, changePct: number) {
   return Math.round(totalAmount - (totalAmount / previousMultiplier))
 }
 
-function getMerchantMarketColor(changePct: number) {
-  if (changePct === 0) {
-    return 'color-mix(in srgb, var(--app-accent) 14%, var(--app-input-bg))'
-  }
-  const variable = changePct < 0 ? 'var(--app-chart-positive)' : 'var(--app-chart-negative)'
-  const mix = Math.min(72, 24 + Math.abs(changePct) * 2.2)
-  return `color-mix(in srgb, ${variable} ${mix}%, var(--app-input-bg))`
-}
-
-function getMerchantTileColor(merchant: MerchantMarketTile) {
-  if (merchant.id === 'other-merchants') {
-    return 'color-mix(in srgb, var(--app-text-muted) 24%, var(--app-input-bg))'
-  }
-  return getMerchantMarketColor(merchant.changePct)
-}
-
 function splitTreemapItems(
   items: MerchantMarketTile[],
   x: number,
@@ -1085,126 +1058,6 @@ function SectionHeader({
       </div>
       <span className="app-label">{label}</span>
       {action && <div className="ml-auto">{action}</div>}
-    </div>
-  )
-}
-
-function MerchantMarketMap({
-  merchants,
-  currency,
-}: {
-  merchants: MerchantMarketTile[]
-  currency: string
-}) {
-  const [hoveredTile, setHoveredTile] = useState<MerchantMarketHover | null>(null)
-
-  return (
-    <div className="relative min-h-0 flex-1">
-      <div className="h-full overflow-hidden rounded-lg border border-[var(--app-border)]">
-        <svg
-          viewBox="0 0 1000 460"
-          preserveAspectRatio="none"
-          role="img"
-          aria-label="Merchant market map"
-          className="h-full w-full"
-        >
-          {merchants.map((merchant) => {
-            const area = merchant.width * merchant.height
-            const labelSize = area > 90000 ? 24 : area > 42000 ? 17 : 12
-            const amountSize = Math.max(labelSize - 5, 10)
-            const amountText = formatCurrency(merchant.totalAmount, currency)
-            return (
-              <g
-                key={merchant.id}
-                onMouseEnter={(event) => {
-                  const rect = event.currentTarget.ownerSVGElement?.getBoundingClientRect()
-                  if (!rect) return
-                  setHoveredTile({
-                    merchant,
-                    x: event.clientX - rect.left,
-                    y: event.clientY - rect.top,
-                  })
-                }}
-                onMouseMove={(event) => {
-                  const rect = event.currentTarget.ownerSVGElement?.getBoundingClientRect()
-                  if (!rect) return
-                  setHoveredTile({
-                    merchant,
-                    x: event.clientX - rect.left,
-                    y: event.clientY - rect.top,
-                  })
-                }}
-                onMouseLeave={() => setHoveredTile(null)}
-              >
-                <rect
-                  x={merchant.x + 2}
-                  y={merchant.y + 2}
-                  width={Math.max(merchant.width - 4, 0)}
-                  height={Math.max(merchant.height - 4, 0)}
-                  rx={6}
-                  fill={getMerchantTileColor(merchant)}
-                  stroke="var(--app-surface-soft)"
-                  strokeWidth={4}
-                />
-                <foreignObject
-                  x={merchant.x + 10}
-                  y={merchant.y + 10}
-                  width={Math.max(merchant.width - 20, 0)}
-                  height={Math.max(merchant.height - 20, 0)}
-                >
-                  <div
-                    className="flex h-full min-w-0 flex-col items-center justify-center text-center"
-                    style={{ color: 'var(--app-text)' }}
-                  >
-                    <p
-                      className="max-w-full break-words font-bold leading-tight"
-                      style={{ fontSize: labelSize }}
-                    >
-                      {merchant.name}
-                    </p>
-                    <p
-                      className="mt-1 max-w-full break-words font-financial leading-tight"
-                      style={{ color: 'var(--app-text-muted)', fontSize: amountSize }}
-                    >
-                      {amountText}
-                    </p>
-                  </div>
-                </foreignObject>
-              </g>
-            )
-          })}
-        </svg>
-      </div>
-      {hoveredTile && (
-        <div
-          className="app-chart-tooltip-default-content pointer-events-none absolute z-20 min-w-56"
-          style={{
-            left: hoveredTile.x,
-            top: hoveredTile.y,
-            transform: 'translate(-50%, calc(-100% - 10px))',
-          }}
-        >
-          <p className="app-tooltip-muted">{hoveredTile.merchant.name}</p>
-          <div className="mt-1 flex justify-between gap-4">
-            <span>Total Spend</span>
-            <span className="font-financial">{formatCurrency(hoveredTile.merchant.totalAmount, currency)}</span>
-          </div>
-          {hoveredTile.merchant.changeAmount === null ? (
-            <p className="mt-1 text-xs" style={{ color: 'var(--app-text-muted)' }}>
-              Change not shown because this group changes by period.
-            </p>
-          ) : (
-            <div className="mt-1 flex justify-between gap-4">
-              <span>Change</span>
-              <span className="font-financial">
-                {formatSignedCurrency(hoveredTile.merchant.changeAmount, currency)}
-                {' '}
-                ({hoveredTile.merchant.changePct > 0 ? '+' : ''}{hoveredTile.merchant.changePct}%)
-              </span>
-            </div>
-          )}
-        </div>
-      )}
     </div>
   )
 }
@@ -1439,27 +1292,11 @@ export default function InsightsPage() {
         />
 
         <section className="grid grid-cols-1 gap-4 min-[1180px]:grid-cols-[minmax(0,1fr)_360px]">
-          <div className="app-card flex min-h-[560px] flex-col">
-            <SectionHeader icon={Store} label="Spending Distribution by Merchant" />
-            <div className="mb-3 flex flex-wrap items-center justify-between gap-3 text-xs" style={{ color: 'var(--app-text-muted)' }}>
-              <span>Tile size shows total spend. Color shows change vs. the comparable period.</span>
-              <div className="flex flex-wrap items-center gap-3">
-                <span className="inline-flex items-center gap-1.5">
-                  <span className="h-2.5 w-2.5 rounded-sm" style={{ background: 'var(--app-chart-positive)' }} />
-                  Spend down
-                </span>
-                <span className="inline-flex items-center gap-1.5">
-                  <span className="h-2.5 w-2.5 rounded-sm" style={{ background: 'var(--app-chart-negative)' }} />
-                  Spend up
-                </span>
-                <span className="inline-flex items-center gap-1.5">
-                  <span className="h-2.5 w-2.5 rounded-sm" style={{ background: 'var(--app-accent)' }} />
-                  Flat
-                </span>
-              </div>
-            </div>
-            <MerchantMarketMap merchants={merchantMarketLayout} currency={displayCurrency} />
-          </div>
+          <MerchantDistributionCard
+            header={<SectionHeader icon={Store} label="Spending Distribution by Merchant" />}
+            merchants={merchantMarketLayout}
+            currency={displayCurrency}
+          />
 
           <div className="app-card">
             <SectionHeader icon={ListChecks} label="Merchant Ranking" />
